create Rating component to replace ui util
