+++ components/skills/SkillsGrid.tsx
@@ -1,6 +1,7 @@
 import { CloudinaryImage } from '@/components/image';
-import { generateStars, truncateString } from '@/app/utils';
+import { Rating } from '@/components/common';
 import { BaseSkill } from '@/types/Skill';
+import { truncateString } from '@/app/utils';
 
 interface SkillsGridProps {
    skills: BaseSkill[];
@@ -17,23 +18,30 @@ export default function SkillsGrid({
       <div className="grid gap-4 sm:grid-cols-2 md:grid-cols-3 lg:grid-cols-4">
          {skills.map((skill, index) => {
             const { cloudinary, name, rating, years } = skill;
+            const { newString: truncatedName } = truncateString(name, 14);
 
             return (
                <div
                   key={index}
                   className="flex items-center justify-center gap-4 rounded-lg border p-4 shadow-md hover:border-bb-teal hover:bg-bb-gray-800 md:cursor-pointer"
                   onClick={() => handleItemClick(name)}>
-                  <div className="relative h-12 w-12">
-                     <CloudinaryImage
-                        cloudinaryId={cloudinary as string}
-                        alt={name}
-                        width={48}
-                        height={48}
-                     />
-                  </div>
+                  {cloudinary && (
+                     <div className="relative h-12 w-12">
+                        <CloudinaryImage
+                           cloudinaryId={cloudinary}
+                           alt={name}
+                           width={48}
+                           height={48}
+                        />
+                     </div>
+                  )}
                   <div>
-                     <div>{truncateString(name, 15).newString}</div>
-                     <div className="flex h-4">{generateStars(rating)}</div>
+                     <div>{truncatedName}</div>
+                     {rating && (
+                        <div className="flex h-4">
+                           <Rating rating={rating} />
+                        </div>
+                     )}
                      <div className="min-w-24 whitespace-nowrap text-[0.7rem] md:text-sm">
                         {years}+ years
                      </div>
